ssh-add: Start a new ssh-agent instance, independently of gnome-keyring
Login to a server requiring an encrypted key, and the password
will only be stored in this new ssh-agent instance, within this
shell. The instance will exit when the shell exits, leaving the
encrypted key effectively locked.

This does *not* make the key secure within your current user space.
An attacker with your user rights will be able to use the unlocked
key via the socket — but it's better than leaving it unlocked for
the duration of your login session.